Fix: Implement testing and ESLint fixes
Implements the comprehensive plan to address test failures and ESLint issues. This includes fixing test mocks for `useAuth`, resolving `any` types, removing unused variables and imports, and separating context hooks to improve Fast Refresh performance. New type interfaces for images, cache stats, and sync status have also been created.
